Improved Decision-Maker
The Decision-Maker checks now if it is possible to execute Actions
based on the number of VNFCInstances that are already deployed. So
in case that the all the VDUs of a specific VNFR reached the minimum
number of VNFCInstances (1 for each) no scaling-in action will be
sent to the executor. The same for scaling-out if the maximum number
is already reached for all the VDUs. These actions are filtered.